move preference related to to dedicated preference package
