+++ app/src/main/java/com/keylesspalace/tusky/di/ActivitiesModule.kt
@@ -18,6 +18,7 @@ package com.keylesspalace.tusky.di
 import com.keylesspalace.tusky.*
 import com.keylesspalace.tusky.components.compose.ComposeActivity
 import com.keylesspalace.tusky.components.instancemute.InstanceListActivity
+import com.keylesspalace.tusky.components.preference.PreferencesActivity
 import com.keylesspalace.tusky.components.report.ReportActivity
 import com.keylesspalace.tusky.components.scheduled.ScheduledTootActivity
 import com.keylesspalace.tusky.components.search.SearchActivity
